requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.25
sqlalchemy2-stubs==0.0.2a16
sqlmodel==0.0.4
# # Editable Git install with no remote (src==1.0)
# -e c:\users\ziadm\desktop\fastapi\projects\sqlmodel
# starlette==0.14.2
# typing-extensions==3.10.0.2
# uvicorn==0.15.0
